Moved benchmarks to Perf.Half.cs
