+++ src/benchmarks/micro/MicroBenchmarks.csproj
@@ -147,8 +147,7 @@
     <Compile Remove="libraries\System.Text.Json\Serializer\WritePreservedReferences.cs" />
     <Compile Remove="libraries\System.Net.Security\SslStreamTests.Context.cs" />
     <Compile Remove="libraries\System.Formats.Cbor\*.cs" />
-    <Compile Remove="libraries\System\ConvertHalfToSingle.cs" />
-    <Compile Remove="libraries\System\ConvertSingleToHalf.cs" />
+    <Compile Remove="libraries\System.Runtime\Perf.Half.cs" />
   </ItemGroup>
 
   <ItemGroup Condition=" '$(TargetFrameworkIdentifier)' == '.NETFramework' Or ('$(TargetFrameworkIdentifier)' == '.NETCoreApp' And '$(_TargetFrameworkVersionWithoutV)' &lt; '6.0')">
@@ -215,5 +214,4 @@
     <!-- Workaround https://github.com/dotnet/project-system/issues/935 -->
     <None Include="**/*.cs" />
   </ItemGroup>
-
 </Project>
